requirements.txt
pushbullet.py==0.11.0
requests==2.18.1
beautifulsoup4==4.6.0
